Test serializing `SystemContractCode`
